Ashanti's BSearch Code
I set up the function so that it only works for lists that aren't null; otherwise it skips to the end at prints -1. To stop it from going on forever, should it get close to the answer but not quite, I only allowed my while loop to perform as many times as there are iterations in the list.